Do not start a timer for blocked token buckets
Starting a timer for blocked token buckets isn't just a waist of
resources, but in case of speed traps created from templates with
`delete_when_full` option it may actually prevent blocking the speed
trap as intended by the user.

It is a realistic use case that to deal with a system overload the
user won't update the templates, but selectively block some speed
traps created from them. However blocked speed traps don't take tokens
from their buckets, so the buckets would get full and then
deleted. After which the next attempt to pass the speed trap would
re-create the bucket unblocked.

The main consequence of allowing a speed trap to exist without an
associated timer is that `speed_trap_token_bucket` should use the
presence of an id in the `ets` table as the source of truth on whether
a speed trap exists or not. But the change also allowed unifying the
code paths for initialising a new speed trap and modifying an existing
one, with the slight (documented) change in the `speed_trap:modify/2`
API that it now always refills the bucket to its maximum capacity.

Finally, this commit cleans a slight naming inconsistency in
`speed_trap_token_bucket`: the variables with `token_bucket()` type
have been renamed from `Ctr` to `Bucket`.